ArchonTestSearchAgent/requirements.txt
pydantic-ai>=0.0.22
pydantic>=2.0.0
requests>=2.0.0
python-dotenv>=1.0.0